renamed to MemorySize

+++ src/data_type/type_modifier.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, PartialEq)]
 pub enum DeclModifier {
     POINTER,//this declaration is for a pointer to something
-    ARRAY(usize),//an array with usize elements
+    ARRAY(u64),//an array with usize elements
 }
